+++ changelog.md
@@ -18,6 +18,7 @@ Added option for restoring `/rc` to its ready check functionality (#215, Curse#4
 - *Fixed occasional missing data error. (Curse#503)*
 - *Fixed hyperlink formats in TSV exports - thanks to Yttrium-tYcLief (#214).*
 - *Added `id` to json export (Classic#43).*
+- *Award Later items now remembers which boss they where dropped by (Classic#43).*
 
 
 # v3.3.0